Remove logic for conditional migrations
This patch removes the logic for handling conditional migrations
and adjusts all the patches where it was used accordingly.

The alembic environment has been update to not send anymore
the active plugins list to migrations.

This patch also removes the vestigial 'options' parameter which
was sent to every migration but always set to None and never
used by any migration.

Implements blueprint bp/reorganize-migrations

+++ neutron/db/migration/alembic_migrations/versions/50e86cb2637a_nsx_mappings.py
@@ -25,25 +25,13 @@ Create Date: 2013-10-26 14:37:30.012149
 revision = '50e86cb2637a'
 down_revision = '1fcfc149aca4'
 
-# Change to ['*'] if this migration applies to all plugins
-
-migration_for_plugins = [
-    'neutron.plugins.nicira.NeutronPlugin.NvpPluginV2',
-    'neutron.plugins.nicira.NeutronServicePlugin.NvpAdvancedPlugin',
-    'neutron.plugins.vmware.plugin.NsxPlugin',
-    'neutron.plugins.vmware.plugin.NsxServicePlugin'
-]
-
 from alembic import op
 import sqlalchemy as sa
 
 from neutron.db import migration
 
 
-def upgrade(active_plugins=None, options=None):
-    if not migration.should_run(active_plugins, migration_for_plugins):
-        return
-
+def upgrade():
     op.create_table('neutron_nsx_port_mappings',
                     sa.Column('neutron_id', sa.String(length=36),
                               nullable=False),
@@ -55,11 +43,13 @@ def upgrade(active_plugins=None, options=None):
                                             ondelete='CASCADE'),
                     sa.PrimaryKeyConstraint('neutron_id'))
 
-    op.execute("INSERT INTO neutron_nsx_port_mappings SELECT quantum_id as "
-               "neutron_id, nvp_id as nsx_port_id, null as nsx_switch_id from"
-               " quantum_nvp_port_mapping")
-    op.drop_table('quantum_nvp_port_mapping')
+    if migration.schema_has_table('quantum_nvp_port_mapping'):
+        op.execute(
+            "INSERT INTO neutron_nsx_port_mappings SELECT quantum_id as "
+            "neutron_id, nvp_id as nsx_port_id, null as nsx_switch_id from"
+            " quantum_nvp_port_mapping")
+        op.drop_table('quantum_nvp_port_mapping')
 
 
-def downgrade(active_plugins=None, options=None):
+def downgrade():
     pass
